clean up training tools
